Changed Vault login strategy to AppRole instead of Token.
AppRole is the far superior auth backend. It is way more suitable for applications like Jenkins.
We now need to supply a RoleID and an SecretID. Based on this credentials we can obtain a token from Vault.
This token is only valid for a couple of minutes then.

We can restrict the RoleID/SecretID-tuple to have a ttl (and all this other stuff). What I recommend, though, is restricting the usage of the AppRole backend via `bound_cidr_list`. This way the RoleID/SecretID-tuple can only be exchanged for a valid access token when the request comes from a certain IP range.